+++ CHANGELOG.md
@@ -5,3 +5,6 @@
 ### Fixes
 - Better retro-compatibility for animations
 - Can't access options screen from main screen in 1.21.1
+
+### Misc
+- Send errors in chat if animations are not properly initialized on cutscene editor